Tests

+++ libft/lib_data/l_gnrl/src/size.c
@@ -6,7 +6,7 @@
 /*   By: eslamber <eslamber@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/14 14:31:44 by eslamber          #+#    #+#             */
-/*   Updated: 2023/01/05 18:50:42 by eslamber         ###   ########.fr       */
+/*   Updated: 2023/03/02 16:17:58 by eslamber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,26 +61,26 @@ void	debugger_size(int debug, t_type t, int mod)
 	}
 }
 
-long long int	size(t_type t, int debug)
+size_t	size(t_type t, int debug)
 {
 	debugger_size(debug, t, 0);
 	debugger_size(debug, t, 1);
 	if (t == INT)
-		return ((long long int)(sizeof(int)));
+		return (sizeof(int));
 	else if (t == CHAR)
-		return ((long long int)(sizeof(char)));
+		return (sizeof(char));
 	else if (t == DOUBLE)
-		return ((long long int)(sizeof(double)));
+		return (sizeof(double));
 	else if (t == LONG)
-		return ((long long int)(sizeof(long)));
+		return (sizeof(long));
 	else if (t == UNSIGNED)
-		return ((long long int)(sizeof(unsigned)));
+		return (sizeof(unsigned));
 	else if (t == LONG_LONG)
-		return ((long long int)(sizeof(long long)));
+		return (sizeof(long long));
 	else if (t == LONG_UNSIGNED)
-		return ((long long int)(sizeof(long unsigned)));
+		return (sizeof(long unsigned));
 	else if (t == LONG_LONG_UNSIGNED)
-		return ((long long int)(sizeof(long long unsigned)));
+		return (sizeof(long long unsigned));
 	else
 		return (-1);
 }
